+++ src/state.rs
@@ -1,8 +1,9 @@
-use std::fmt::Debug;
+use std::fmt::{Debug, Formatter, Result};
 
-use bevy::ecs::system::EntityCommands;
-
-use crate::prelude::*;
+use crate::{
+    bundle::{Insert, Remove},
+    prelude::*,
+};
 
 /// Represents a state that an entity may be in. A state must implement [`Reflect`],
 /// but workarounds exist for structs that contain types that do not implement [`Reflect`]
@@ -45,33 +46,23 @@ use crate::prelude::*;
 ///
 /// These workarounds currently do not affect the functionality of your state machine.
 ///
-/// If you are concerned with performance, I recommend having your states use sparse set storage,
+/// If you are concerned with performance, consider having your states use sparse set storage,
 /// since they may be added to and removed from entities.
 pub trait MachineState: 'static + Bundle + Clone + Reflect + Send + Sync {}
 
 impl<T: 'static + Bundle + Clone + Reflect + Send + Sync> MachineState for T {}
 
-pub(crate) trait StateTransition: 'static + Reflect + Send + Sync {
-    fn insert(&self, entity: &mut EntityCommands);
-    fn remove(&self, entity: &mut EntityCommands);
+pub(crate) trait StateTransition: 'static + Insert + Reflect + Remove + Send + Sync {
     fn dyn_clone(&self) -> Box<dyn StateTransition>;
 }
 
 impl Debug for dyn StateTransition {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result {
         self.as_reflect().fmt(f)
     }
 }
 
 impl<T: MachineState> StateTransition for T {
-    fn insert(&self, entity: &mut EntityCommands) {
-        entity.insert_bundle(self.clone());
-    }
-
-    fn remove(&self, entity: &mut EntityCommands) {
-        entity.remove_bundle::<T>();
-    }
-
     fn dyn_clone(&self) -> Box<dyn StateTransition> {
         Box::new(self.clone())
     }